:sparkles:[Feat] #S11P12A106-130 #done 마이페이지 구현 완료

See merge request s11-webmobile1-sub2/S11P12A106!86

+++ back/src/main/java/com/ssafy/ditto/domain/mypage/dto/LikeClassResponse.java
@@ -11,8 +11,8 @@ public class LikeClassResponse {
     private Integer classId;
     private String className;
     private Integer classPrice;
-    private Integer classHour;
-    private Integer classMinute;
+    private Byte classHour;
+    private Byte classMinute;
     private Integer likeCount;
     private Integer reviewCount;
     private Integer ratingSum;
